fix: address PR review comments
