+++ components/input/src/i18n.js
@@ -42,7 +42,7 @@ const stringsES = {
   'validCard': 'Por favor, introduzca un número de tarjeta de crédito válido.',
   'email': 'Introduzca una dirección de correo electrónico válida (nombre@dominio.com).',
   'password': `Las contraseñas válidas deben constar de al menos 8 caracteres, incluyendo al menos una letra mayúscula, una letra minúscula y un número.`,
-  'creditcard': 'Por favor, introduzca un número de tarjeta de crédito válido.',
+  'credit-card': 'Por favor, introduzca un número de tarjeta de crédito válido.',
   'tel': 'Por favor, introduzca un número de teléfono válido.',
   'dateYYYY': 'Ingrese una fecha completa en el formato AAAA',
   'dateYYYYMM': 'Ingrese una fecha completa en el formato AAAA/MM',
@@ -65,7 +65,7 @@ const stringsEN = {
   'validCard': 'Please enter a valid credit card number.',
   'email': 'Please enter a valid email address (name@domain.com).',
   'password': 'Valid passwords must consist of at least 8 characters, including at least one uppercase letter, one lowercase letter, and one number.',
-  'creditcard': 'Please enter a valid credit card number.',
+  'credit-card': 'Please enter a valid credit card number.',
   'tel': 'Please enter a valid phone number.',
   'dateYYYY': 'Please enter a complete date in the format YYYY',
   'dateYYYYMM': 'Please enter a complete date in the format YYYY/MM',
